# CASSANDRA-19015 Nodetool 'tablestats' formatting uses inconsistent significant digits