--- a/test/unit/org/apache/cassandra/utils/FBUtilitiesTest.java
+++ b/test/unit/org/apache/cassandra/utils/FBUtilitiesTest.java
@@ -392,10 +392,10 @@ public class FBUtilitiesTest
     @Test
     public void testPrettyPrintAverage()
     {
-        Assert.assertEquals("100500.0", FBUtilities.prettyPrintAverage(100500));
-        Assert.assertEquals("1.5", FBUtilities.prettyPrintAverage(1.5));
+        Assert.assertEquals("100500.00", FBUtilities.prettyPrintAverage(100500));
+        Assert.assertEquals("1.50", FBUtilities.prettyPrintAverage(1.5));
         Assert.assertEquals("0.05", FBUtilities.prettyPrintAverage(0.05));
-        Assert.assertEquals("0.0", FBUtilities.prettyPrintAverage(0.00));
+        Assert.assertEquals("0.00", FBUtilities.prettyPrintAverage(0.00));
         Assert.assertEquals("NaN", FBUtilities.prettyPrintAverage(Double.NaN));
         Assert.assertEquals("Infinity", FBUtilities.prettyPrintAverage(Double.POSITIVE_INFINITY));
     }