… test/unit/org/apache/cassandra/tools/nodetool/stats/TableStatsPrinterTest.java